% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot.BasicRaster}
\alias{ggplot.BasicRaster}
\title{ggplot for raster, fast and easy}
\usage{
ggplot.BasicRaster(
  data = NULL,
  mapping = aes(),
  ...,
  environment = parent.frame()
)
}
\arguments{
\item{data}{BasicRaster (raster, stack, or brick)}

\item{mapping}{don't use this, use aes() down stream for adding layers}

\item{...}{pass arguments to [plus_raster()]}

\item{environment}{don't use this either}
}
\value{
a [ggplot()] object
}
\description{
A [ggplot()] method for raster objects, it plots them as-is, like [image()]
for a single layer, and like [raster::plotRGB()] for three or four layers (assuming type byte
effectively. Data is automatically scaled from what it is to '[0,1]' internally.
}
\examples{
library(ggplot2)
sfx <- sf::st_as_sf(data.frame(x = rnorm(100), y = rnorm(100)), coords = c("x", "y"))
gp <- geom_point(data = data.frame(x = runif(5), y = runif(5)), aes(x, y), col = "hotpink", pch = "+", cex = 20)
ggplot(raster::raster(volcano)) +  gp + geom_sf(data = sfx)
pts <- c(147.3, 147.35, -42.89, -42.87)
cc <- ceramic::cc_location(raster::extent(pts), zoom = 15)
sfx <- sf::st_transform(sf::st_sfc(sf::st_multipoint(matrix(pts, ncol = 2)), crs = 4326), raster::projection(cc))
ggplot(cc) + coord_equal() + geom_sf(data = sfx)  ## it's WROKING
}
